Add files via upload
The Verilog files containing the CGRA implementation are found in the following filepath:

.../PE-4x4-vck190_will/PE-4x4-vck190.srcs/sources_1/imports/Desktop/research/clean/modules/*

To import into Vivado, extract the project from the zip file and create a new project. Then you will need to convert the .v file syntax to SystemVerilog by selecting all of the sources and editing their "Source File Properties." The type will need to be changed from Verilog -> SystemVerilog.